cleaned up all search bar files and components. I DID IT ;_;
